fix status-line to play nice with the legacy loader
