+++ src/extensions/reporter/status-line.ts
@@ -1,8 +1,8 @@
 import stc from 'string-to-color';
 import chalk from 'chalk';
-import ora from 'ora';
 import debounce from 'debounce';
 import getColumnCount from './get-column-count';
+import loader from '../../cli/loader';
 
 // this number is added to status line length calculations.
 // the idea is to assume we have a longer status line to make
@@ -20,56 +20,15 @@ function clearStatusRow() {
   console.log(`\r${Array(getColumnCount()).fill(' ').join('')}`);
 }
 
-// we send a proxy to the spinner instance rather than process.stdout
-// so that we would be able to bypass our monkey-patch of process.stdout
-// this is so that we won't have a case where the stdout "write" method
-// triggers itself through the spinner by doing "spinner.start()" or "spinner.stop()"
-const originalStdoutWrite = process.stdout.write.bind(process.stdout);
-const stdoutProxy = new Proxy(process.stdout, {
-  get(obj, prop) {
-    if (prop === 'write') {
-      return originalStdoutWrite;
-    }
-    return obj[prop];
-  },
-});
-
-const originalStderrWrite = process.stderr.write.bind(process.stderr);
-
 export default class StatusLine {
   public buffer = SPACE_BUFFER;
-  private spinner: any = ora({ spinner: 'bouncingBar', stream: stdoutProxy, isEnabled: true }).stop();
+  private spinner = loader;
   private spinnerLength = 7; // 6 for spinner, 1 for space after it
   private text = '';
   private ids: Array<string> = [];
   private ended = false;
   constructor() {
     this.reRender = debounce(this.reRender, TIME_BETWEEN_RE_RENDERING);
-    // @ts-ignore
-    // here we monkey-patch the process.stdout stream so that whatever is printed
-    // does not break the status line with the spinner, and that this line always
-    // remains at the bottom of the screen
-    process.stdout.write = (buffer, encoding, callback) => {
-      const wasSpinning = this.spinner.isSpinning;
-      if (wasSpinning) {
-        this.spinner.stop();
-      }
-      originalStdoutWrite(buffer, encoding, callback);
-      if (wasSpinning) {
-        this.spinner.start();
-      }
-    };
-    // @ts-ignore
-    process.stderr.write = (buffer, encoding, callback) => {
-      const wasSpinning = this.spinner.isSpinning;
-      if (wasSpinning) {
-        this.spinner.stop();
-      }
-      originalStderrWrite(buffer, encoding, callback);
-      if (wasSpinning) {
-        this.spinner.start();
-      }
-    };
   }
   addId(id) {
     this.ids.push(id);
@@ -119,13 +78,13 @@ export default class StatusLine {
     if (columnCount < spinnerLength + 10) {
       clearStatusRow();
     } else if (columnCount < this.shortStatusLine(this.text).length + spinnerLength + SPACE_BUFFER) {
-      this.spinner.text = this.text;
+      this.spinner.setText(this.text);
       this.spinner.start();
     } else if (columnCount < this.fullVersionLength(this.text) + this.spinnerLength + SPACE_BUFFER) {
-      this.spinner.text = this.shortStatusLine(this.text);
+      this.spinner.setText(this.shortStatusLine(this.text));
       this.spinner.start();
     } else {
-      this.spinner.text = this.fullVersion(this.text);
+      this.spinner.setText(this.fullVersion(this.text));
       this.spinner.start();
     }
   }
